improve rdoc

+++ lib/tinyci/compactor.rb
@@ -6,7 +6,7 @@ require 'find'
 
 module TinyCI
   
-  # Responsible for writing the git hook file
+  # Tool for compressing old builds into .tar.gz files
   class Compactor
     include GitUtils
     include Subprocesses
@@ -15,18 +15,25 @@ module TinyCI
     
     # Constructor
     # 
-    # @param [String] working_dir The directory from which to run. Does not have to be the root of the repo
+    # @param [String] working_dir The directory from which to run.
+    # @param [Integer] num_builds_to_leave How many builds not to compact, starting from the newest
+    # @param [String] builds_to_leave Comma-separated list of build directory names not to compact
     # @param [Logger] logger Logger object
     def initialize(working_dir: nil, num_builds_to_leave: nil, builds_to_leave: nil, logger: nil)
       @logger = logger
       @working_dir = working_dir || repo_root
-      @num_builds_to_leave = num_builds_to_leave || 1
+      @num_builds_to_leave = (num_builds_to_leave || 1).to_i
       @builds_to_leave = builds_to_leave || []
     end
     
-    # Write the hook to the relevant path and make it executable
+    # Compress and delete the build directories
     def compact!
-      directories_to_compress.each do |dir|
+      unless inside_repository?
+        log_error "not currently in a git repository"
+        return false
+      end
+      
+      directories_to_compact.each do |dir|
         compress_directory dir
         FileUtils.rm_rf builds_dir(dir)
         
@@ -36,7 +43,8 @@ module TinyCI
     
     private
     
-    def directories_to_compress
+    # Build the list of directories to compact according to the options
+    def directories_to_compact
       builds = Dir.entries builds_dir
       builds.select! {|e| File.directory? builds_dir(e) }
       builds.reject! {|e| %w{. ..}.include? e }
@@ -48,14 +56,18 @@ module TinyCI
       builds
     end
     
+    # Get the location of the builds directory
     def builds_dir(path_segments = nil)
       File.join @working_dir, 'builds/', *path_segments
     end
     
+    # Build the path for a compressed archive
     def archive_path(dir)
       File.join(builds_dir, dir+".tar.gz")
     end
     
+    # Create a .tar.gz file from a directory
+    # Done in pure ruby to ensure portability
     def compress_directory(dir)
       File.open archive_path(dir), 'wb' do |oarchive_path|
         Zlib::GzipWriter.wrap oarchive_path do |gz|
